chore: fix the issue of the refiner_Switches at param being invalid in the Novita.AI tool, and modularize the handling of its complex params.
